requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
psycopg2-binary==2.9.9
Werkzeug==3.0.1
gunicorn==21.2.0
SQLAlchemy<2.0
Flask-WTF==0.15.1
python-dotenv==1.0.0
email-validator==1.1.3
Flask-Mail==0.9.1
# Add your other project dependencies here 